wis2box-create-config initial version (#476)
* wis2box-create-config initial version

* flake8 corrections

* add station_list_example.csv

* new setup instructions

* remove setup-new

* use file templating, rename template files, update doc

* fix doc

* update countries JSON, add country name to workflow

---------

+++ wis2box-ctl.py
@@ -162,6 +162,9 @@ def make(args) -> None:
         run(args, split(
             f'{DOCKER_COMPOSE_COMMAND} {docker_compose_args} build {containers}'))
     elif args.command in ["up", "start", "start-dev"]:
+        if not os.path.exists('dev.env'):
+            print("ERROR: dev.env file does not exist.  Please create one manually or by running `python3 wis2box-create-config.py`")
+            exit(1)
         run(args, split(
             'docker plugin install grafana/loki-docker-driver:latest --alias loki --grant-all-permissions > /dev/null 2>&1'))
         run(args, split(
